+++ src/app/[locale]/(groups)/groups/[id]/layout.tsx
@@ -1,28 +1,6 @@
 import Banner from '@/components/Banner/Banner';
 import DivisionGroupService from '@/services/divisionGroupService';
 import GammaService from '@/services/gammaService';
-import { Metadata } from 'next';
-import i18nService from '@/services/i18nService';
-
-export async function generateMetadata(props: {
-  params: Promise<{ locale: string; id: string }>;
-}) {
-  const params = await props.params;
-
-  const { locale, id } = params;
-
-  const group = await DivisionGroupService.getInfoBySlug(id).catch(() => {
-    return null;
-  });
-  const l = i18nService.getLocale(locale);
-  return {
-    title:
-      group !== null
-        ? group.prettyName
-        : undefined,
-    description: l.site.siteDescription
-  } as Metadata;
-}
 
 export default async function GroupLayout(props: {
   children: React.ReactNode;
