+++ src/app/[locale]/(groups)/groups/[id]/page.tsx
@@ -11,8 +11,29 @@ import ContactCard from '@/components/ContactCard/ContactCard';
 import i18nService from '@/services/i18nService';
 import ActionLink from '@/components/ActionButton/ActionLink';
 import GroupAvatar from '@/components/GroupAvatar/GroupAvatar';
+import { Metadata } from 'next';
 
 export const revalidate = 3600;
+
+export async function generateMetadata(props: {
+  params: Promise<{ locale: string; id: string }>;
+}) {
+  const params = await props.params;
+
+  const { locale, id } = params;
+
+  const group = await DivisionGroupService.getInfoBySlug(id).catch(() => {
+    return null;
+  });
+  const l = i18nService.getLocale(locale);
+  return {
+    title:
+      group !== null
+        ? group.prettyName
+        : undefined,
+    description: l.site.siteDescription
+  } as Metadata;
+}
 
 export default async function Page(props: {
   params: Promise<{ locale: string; id: string }>;
